split import detection into its own module
